+++ lib/ui/emulator/emulator_widget.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:nesd/ui/emulator/debug_overlay.dart';
 import 'package:nesd/ui/emulator/display.dart';
 import 'package:nesd/ui/emulator/input/touch/touch_controls.dart';
 import 'package:nesd/ui/settings/settings.dart';
@@ -14,6 +15,7 @@ class EmulatorWidget extends ConsumerWidget {
     return Stack(
       children: [
         const DisplayWidget(),
+        if (settings.showDebugOverlay) const DebugOverlay(),
         if (settings.showTouchControls) const TouchControls(),
       ],
     );