+++ lib/ui/settings/debug/debug_settings.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:nesd/ui/settings/debug/cartridge_switch.dart';
+import 'package:nesd/ui/settings/debug/debug_overlay_switch.dart';
 import 'package:nesd/ui/settings/debug/debug_tile_switch.dart';
 import 'package:nesd/ui/settings/settings_tab.dart';
 
@@ -14,6 +15,7 @@ class DebugSettings extends StatelessWidget {
         children: const [
           DebugTileSwitch(),
           CartridgeSwitch(),
+          DebugOverlaySwitch(),
         ],
       ),
     );