+++ lib/ui/toast/toast_overlay.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:nesd/ui/common/outline_text.dart';
 import 'package:nesd/ui/nesd_theme.dart';
 import 'package:nesd/ui/toast/toaster.dart';
 
@@ -51,8 +52,6 @@ class ToastWidget extends ConsumerWidget {
       fontWeight: FontWeight.bold,
     );
 
-    const blurRadius = 0.5;
-
     final outlineColor = switch (toast.type) {
       ToastType.info => Colors.blue[100]!,
       ToastType.warning => Colors.orange[100]!,
@@ -66,33 +65,16 @@ class ToastWidget extends ConsumerWidget {
         child: Container(
           margin: const EdgeInsets.only(bottom: 8),
           height: 24,
-          child: Text(
+          child: OutlineText(
             toast.message,
+            outlineColor: outlineColor,
+            blurRadius: 0.5,
             style: textStyle.copyWith(
               color: switch (toast.type) {
                 ToastType.info => Colors.blue[700],
                 ToastType.warning => Colors.orange[800],
                 ToastType.error => nesdRed[600],
               },
-              // ugly hack for outlined text
-              shadows: [
-                Shadow(
-                  color: outlineColor,
-                  blurRadius: blurRadius,
-                ),
-                Shadow(
-                  color: outlineColor,
-                  blurRadius: blurRadius,
-                ),
-                Shadow(
-                  color: outlineColor,
-                  blurRadius: blurRadius,
-                ),
-                Shadow(
-                  color: outlineColor,
-                  blurRadius: blurRadius,
-                ),
-              ],
             ),
           ),
         ),